Fixed bugs + Ui changes

+++ ProjectMayhem/ProjectMayhem/Controllers/AccountController.cs
@@ -279,10 +279,12 @@ namespace ProjectMayhem.Controllers
             Debug.WriteLine(SelectedId + " This ID");
             var currentUser = User.Identity.GetUserId();
             var deletionUser = UserManager.FindById(SelectedId);
-            if (UserManager.Users.Where(x => x.teamLead.Id == deletionUser.Id).ToArray().Length == 0 && deletionUser.teamLead.Id == currentUser)
+            if (UserManager.Users.Where(x => x.teamLead.Id == deletionUser.Id).ToArray().Length == 0 && deletionUser.teamLead.Id == currentUser && deletionUser != null)
             {
                 await UserManager.DeleteAsync(deletionUser);
             }
+            else if (deletionUser == null)
+                TempData["DeleteError"] = "The Team member does not exist";
             else
                 TempData["DeleteError"] = "Cannot delete a member. User is a Team Leader or you do not have permission";
             return this.RedirectToAction("Register");
